backend/requirements.txt

flask==2.0.1
flask-cors==3.0.10
pdf2image==1.16.0
opencv-python==4.5.3.56
pytesseract==0.3.8
psycopg2-binary==2.9.1
numpy==1.21.2
Pillow==8.3.2
